# byclaw-tenant-node/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ByClaw fixed tenant node",
    "version": "1.0.0",
    "description": "BE-only mTLS API. Canonical business hash and Redis/Worker contracts are specified in README. No raw SQL/bootstrap/message overwrite API."
  },
  "servers": [
    {
      "url": "https://tenant-node.example.test:3100"
    }
  ],
  "security": [
    {
      "mtls": []
    }
  ],
  "paths": {
...
    "/internal/v1/schema": {
      "get": {
        "summary": "Observed and audited schema state",
        "parameters": [
          {
            "$ref": "#/components/parameters/enterprise"
          },
          {
            "$ref": "#/components/parameters/generation"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Rejected; sanitized error code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Rejected; sanitized error code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Rejected; sanitized error code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Rejected; sanitized error code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Rejected; sanitized error code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Rejected; sanitized error code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "mtls": {
        "type": "mutualTLS",
        "description": "Trusted BE certificate chain and pinned CN"
      }
    },
    "parameters": {
      "enterprise": {
        "name": "X-Enterprise-Id",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string",
          "pattern": "^[1-9][0-9]{0,18}$",
          "description": "规范十进制 signed bigint ID；禁止 JS Number"
        }
      },
      "generation": {
        "name": "X-Tenant-Generation",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string",
          "pattern": "^[1-9][0-9]{0,18}$",
          "description": "规范十进制 signed bigint ID；禁止 JS Number"
        }
      },
...
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "object",
            "required": ["code"],
            "properties": {
              "code": {
                "type": "string"
              }
            }
          }
        }
      },
...
    }
  }
}